+++ packages/tooling/fast-tooling/src/web-components/html-render/index.ts
@@ -1,10 +1,11 @@
-import { HTMLRender } from "./html-render";
-import { HTMLRenderTemplate as template } from "./html-render.template";
+import { HTMLRender, htmlRenderOriginatorId } from "./html-render";
+import { htmlRenderTemplate as template } from "./html-render.template";
 import { htmlRenderStyles as styles } from "./html-render.styles";
 
 /**
+ * A web component for rendering HTML using the MessageSystem.
  *
- * @public
+ * @alpha
  * @remarks
  * HTML Element: \<html-render\>
  */
@@ -13,3 +14,5 @@ export const fastToolingHTMLRender = HTMLRender.compose({
     template,
     styles,
 });
+
+export { htmlRenderOriginatorId };