+++ packages/web-components/fast-foundation/src/number-field/number-field.ts
@@ -127,6 +127,7 @@ export class NumberField extends FormAssociatedNumberField {
                 this.max = numb;
             }
         }
+        this.updateValue(this.value);
     }
 
     /**
@@ -147,6 +148,7 @@ export class NumberField extends FormAssociatedNumberField {
                 this.min = numb;
             }
         }
+        this.updateValue(this.value);
     }
 
     /**
@@ -169,38 +171,38 @@ export class NumberField extends FormAssociatedNumberField {
     public valueChanged(previousValue, nextValue): void {
         super.valueChanged(previousValue, nextValue);
 
-        const numb = parseFloat(nextValue);
-        let out: number | string = numb == nextValue ? nextValue : numb;
-
-        if (nextValue === "" || isNaN(numb)) {
-            out = "";
-        } else {
-            out = this.getValidValue(numb);
-        }
-
-        this.value = out;
-
-        if (this.proxy instanceof HTMLElement) {
-            this.proxy.value = this.value;
-        }
+        this.updateValue(nextValue);
     }
 
     /**
-     * Ensures that the value is between the min and max values
+     * Updates the value. Validates that it's a number, between the min
+     *  and max, updates the proxy and emits events.
      *
-     * @param value - number to evaluate
-     * @returns - a string repesentation
-     *
+     * @param value - value to be validated
      * @internal
      */
-    private getValidValue(value: number): string {
-        if (this.min !== undefined && value < this.min) {
-            value = this.min;
-        } else if (this.max !== undefined && value > this.max) {
-            value = this.max;
+    private updateValue(value): void {
+        if (value === "" || isNaN(parseFloat(value))) {
+            value = "";
+        } else {
+            value = parseFloat(value);
+            if (this.min !== undefined && value < this.min) {
+                value = this.min;
+            } else if (this.max !== undefined && value > this.max) {
+                value = this.max;
+            }
+
+            value = parseFloat(value.toPrecision(12)).toString();
         }
 
-        return parseFloat(value.toPrecision(12)).toString();
+        if (value != this.value) {
+            this.value = value;
+            if (this.proxy instanceof HTMLInputElement) {
+                this.proxy.value = this.value;
+            }
+            this.$emit("input");
+            this.$emit("change");
+        }
     }
 
     /**
@@ -210,9 +212,7 @@ export class NumberField extends FormAssociatedNumberField {
      */
     public stepUp(): void {
         const stepUpValue = this.step + (parseFloat(this.value) || 0);
-        this.value = this.getValidValue(stepUpValue);
-
-        this.$emit("input");
+        this.updateValue(stepUpValue);
     }
 
     /**
@@ -222,9 +222,7 @@ export class NumberField extends FormAssociatedNumberField {
      */
     public stepDown(): void {
         const stepDownValue = (parseFloat(this.value) || 0) - this.step;
-        this.value = this.getValidValue(stepDownValue);
-
-        this.$emit("input");
+        this.updateValue(stepDownValue);
     }
 
     /**