+++ packages/tooling/fast-figma-plugin-msft/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @microsoft/fast-figma-plugin-msft
 
-This log was last generated on Thu, 15 Jul 2021 01:41:48 GMT and should not be manually modified.
+This log was last generated on Wed, 21 Jul 2021 07:14:48 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.8.16
+
+Wed, 21 Jul 2021 07:14:48 GMT
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.3.0 (roeisenb@microsoft.com)
+
+## 0.8.16
+
+Sun, 18 Jul 2021 07:16:22 GMT
+
+### Patches
+
+- Bump @microsoft/fast-components to v2.2.3 (ben@platform5.ch)
 
 ## 0.8.16